api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
mysqlclient==2.1.1
python-decouple==3.8
